Fixed deleting monitor workflow metadata

+++ alerting/src/main/kotlin/org/opensearch/alerting/transport/TransportDeleteWorkflowAction.kt
@@ -10,6 +10,7 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import org.apache.logging.log4j.LogManager
 import org.apache.lucene.search.join.ScoreMode
+import org.opensearch.OpenSearchException
 import org.opensearch.OpenSearchStatusException
 import org.opensearch.action.ActionListener
 import org.opensearch.action.ActionRequest
@@ -22,6 +23,8 @@ import org.opensearch.action.search.SearchResponse
 import org.opensearch.action.support.ActionFilters
 import org.opensearch.action.support.HandledTransportAction
 import org.opensearch.action.support.WriteRequest.RefreshPolicy
+import org.opensearch.alerting.model.MonitorMetadata
+import org.opensearch.alerting.model.WorkflowMetadata
 import org.opensearch.alerting.opensearchapi.InjectorContextElement
 import org.opensearch.alerting.opensearchapi.addFilter
 import org.opensearch.alerting.opensearchapi.suspendUntil
@@ -43,6 +46,7 @@ import org.opensearch.commons.alerting.action.DeleteMonitorResponse
 import org.opensearch.commons.alerting.action.DeleteWorkflowRequest
 import org.opensearch.commons.alerting.action.DeleteWorkflowResponse
 import org.opensearch.commons.alerting.model.CompositeInput
+import org.opensearch.commons.alerting.model.Monitor
 import org.opensearch.commons.alerting.model.ScheduledJob
 import org.opensearch.commons.alerting.model.Workflow
 import org.opensearch.commons.authuser.User
@@ -51,6 +55,9 @@ import org.opensearch.core.xcontent.NamedXContentRegistry
 import org.opensearch.core.xcontent.XContentParser
 import org.opensearch.index.IndexNotFoundException
 import org.opensearch.index.query.QueryBuilders
+import org.opensearch.index.reindex.BulkByScrollResponse
+import org.opensearch.index.reindex.DeleteByQueryAction
+import org.opensearch.index.reindex.DeleteByQueryRequestBuilder
 import org.opensearch.rest.RestStatus
 import org.opensearch.search.builder.SearchSourceBuilder
 import org.opensearch.tasks.Task
@@ -130,13 +137,13 @@ class TransportDeleteWorkflowAction @Inject constructor(
 
                 if (canDelete) {
                     val delegateMonitorIds = (workflow.inputs[0] as CompositeInput).getMonitorIds()
-
+                    var deletableMonitors = listOf<Monitor>()
                     // User can only delete the delegate monitors only in the case if all monitors can be deleted
                     // Partial monitor deletion is not available
                     if (deleteDelegateMonitors == true) {
-                        val monitorIdsToBeDeleted = getDeletableDelegates(workflowId, delegateMonitorIds, user)
+                        deletableMonitors = getDeletableDelegates(workflowId, delegateMonitorIds, user)
                         val monitorsDiff = delegateMonitorIds.toMutableList()
-                        monitorsDiff.removeAll(monitorIdsToBeDeleted)
+                        monitorsDiff.removeAll(deletableMonitors.map { it.id })
 
                         if (monitorsDiff.isNotEmpty()) {
                             actionListener.onFailure(
@@ -150,7 +157,11 @@ class TransportDeleteWorkflowAction @Inject constructor(
                         }
                     }
 
-                    val deleteResponse = deleteWorkflow(workflow)
+                    val deleteResponse = deleteWorkflow(deleteRequest)
+                    val workflowMetadataId = WorkflowMetadata.getId(workflow.id)
+
+                    val metadataIdsToDelete = mutableListOf(workflowMetadataId)
+
                     if (deleteDelegateMonitors == true) {
                         if (user != null && filterByEnabled) {
                             // Un-stash the context
@@ -168,6 +179,25 @@ class TransportDeleteWorkflowAction @Inject constructor(
                         } else {
                             deleteMonitors(delegateMonitorIds, RefreshPolicy.IMMEDIATE)
                         }
+                        // Delete monitors workflow metadata
+                        // Monitor metadata will be in monitorId-workflowId-metadata format
+                        metadataIdsToDelete.addAll(deletableMonitors.map { MonitorMetadata.getId(it, workflowMetadataId) })
+                    }
+                    // Delete the monitors and it's metadata
+                    val deleteMonitorWorkflowMetadataResponse: BulkByScrollResponse = client.suspendUntil {
+                        DeleteByQueryRequestBuilder(client, DeleteByQueryAction.INSTANCE)
+                            .source(ScheduledJob.SCHEDULED_JOBS_INDEX)
+                            .filter(QueryBuilders.idsQuery().addIds(*metadataIdsToDelete.toTypedArray()))
+                            .execute(it)
+                    }
+
+                    if (deleteMonitorWorkflowMetadataResponse.isTimedOut) {
+                        actionListener.onFailure(
+                            AlertingException.wrap(
+                                OpenSearchException("Cannot determine that the ${ScheduledJob.SCHEDULED_JOBS_INDEX} index is healthy")
+                            )
+                        )
+                        return
                     }
                     actionListener.onResponse(DeleteWorkflowResponse(deleteResponse.id, deleteResponse.version))
                 } else {
@@ -210,7 +240,7 @@ class TransportDeleteWorkflowAction @Inject constructor(
          * @param workflowIdToBeDeleted Id of the workflow that should be deleted
          * @param monitorIds List of delegate monitor ids (underlying monitor ids)
          */
-        private suspend fun getDeletableDelegates(workflowIdToBeDeleted: String, monitorIds: List<String>, user: User?): List<String> {
+        private suspend fun getDeletableDelegates(workflowIdToBeDeleted: String, monitorIds: List<String>, user: User?): List<Monitor> {
             // Retrieve monitors belonging to another workflows
             val queryBuilder = QueryBuilders.boolQuery().mustNot(QueryBuilders.termQuery("_id", workflowIdToBeDeleted)).filter(
                 QueryBuilders.nestedQuery(
@@ -228,11 +258,6 @@ class TransportDeleteWorkflowAction @Inject constructor(
             val searchRequest = SearchRequest()
                 .indices(ScheduledJob.SCHEDULED_JOBS_INDEX)
                 .source(SearchSourceBuilder().query(queryBuilder))
-
-            // Check if user can access the monitors(since the monitors could get modified later and the user might not have the backend roles to access the monitors)
-            if (user != null && filterByEnabled) {
-                addFilter(user, searchRequest.source(), "monitor.user.backend_roles.keyword")
-            }
 
             val searchResponse: SearchResponse = client.suspendUntil { search(searchRequest, it) }
 
@@ -252,7 +277,32 @@ class TransportDeleteWorkflowAction @Inject constructor(
             }
             val workflowMonitors = workflows.filter { it.id != workflowIdToBeDeleted }.flatMap { (it.inputs[0] as CompositeInput).getMonitorIds() }.distinct()
             // Monitors that can be deleted -> all workflow delegates - monitors belonging to different workflows
-            return monitorIds.minus(workflowMonitors.toSet())
+            val deletableMonitorIds = monitorIds.minus(workflowMonitors.toSet())
+
+            val query = QueryBuilders.boolQuery().filter(QueryBuilders.termsQuery("_id", deletableMonitorIds))
+            val searchSource = SearchSourceBuilder().query(query)
+            val monitorSearchRequest = SearchRequest(ScheduledJob.SCHEDULED_JOBS_INDEX).source(searchSource)
+
+            if (user != null && filterByEnabled) {
+                addFilter(user, monitorSearchRequest.source(), "monitor.user.backend_roles.keyword")
+            }
+
+            val searchMonitorResponse: SearchResponse = client.suspendUntil { client.search(monitorSearchRequest, it) }
+            if (searchMonitorResponse.isTimedOut) {
+                throw OpenSearchException("Cannot determine that the ${ScheduledJob.SCHEDULED_JOBS_INDEX} index is healthy")
+            }
+            val deletableMonitors = mutableListOf<Monitor>()
+            for (hit in searchMonitorResponse.hits) {
+                XContentType.JSON.xContent().createParser(
+                    xContentRegistry,
+                    LoggingDeprecationHandler.INSTANCE, hit.sourceAsString
+                ).use { hitsParser ->
+                    val monitor = ScheduledJob.parse(hitsParser, hit.id, hit.version) as Monitor
+                    deletableMonitors.add(monitor)
+                }
+            }
+
+            return deletableMonitors
         }
 
         private suspend fun getWorkflow(): Workflow {
@@ -273,13 +323,13 @@ class TransportDeleteWorkflowAction @Inject constructor(
             return ScheduledJob.parse(xcp, getResponse.id, getResponse.version) as Workflow
         }
 
-        private suspend fun deleteWorkflow(workflow: Workflow): DeleteResponse {
+        private suspend fun deleteWorkflow(deleteRequest: DeleteRequest): DeleteResponse {
             log.debug("Deleting the workflow with id ${deleteRequest.id()}")
             return client.suspendUntil { delete(deleteRequest, it) }
         }
-        // TODO - use once the workflow metadata concept is introduced
-        private suspend fun deleteMetadata(workflow: Workflow) {
-            val deleteRequest = DeleteRequest(ScheduledJob.SCHEDULED_JOBS_INDEX, "${workflow.id}-metadata")
+
+        private suspend fun deleteWorkflowMetadata(workflow: Workflow) {
+            val deleteRequest = DeleteRequest(ScheduledJob.SCHEDULED_JOBS_INDEX, WorkflowMetadata.getId(workflow.id))
             val deleteResponse: DeleteResponse = client.suspendUntil { delete(deleteRequest, it) }
         }
     }
